services/modules/pyproject.toml
[tool.poetry]
name = "portfolio-modules"
version = "0.1.0"
description = ""
authors = ["iizhukov <iizhukov@icloud.com>"]
packages = [
    { include = "core" },
    { include = "models" },
    { include = "scripts" },
    { include = "services" },
    { include = "workers" }
]

[tool.poetry.dependencies]
python = "^3.12"
grpcio = ">=1.70.0"
protobuf = ">=6.0.0"
sqlalchemy = "2.0.44"
asyncpg = "0.30.0"
greenlet = "^3.2.4"

[tool.poetry.dependencies.portfolio-shared]
path = "../../shared/python"
develop = true

[tool.poetry.group.dev.dependencies]
grpcio-tools = ">=1.70.0"
pytest = "8.0.0"
pytest-asyncio = "0.23.0"
mypy-protobuf = "^3.6.0"

[tool.poetry.scripts]
generate-grpc = "scripts.generate_grpc:main"
init-db = "scripts.init_db:main"
service = "scripts.run_service:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
